* Implemented MVI for: Category List and Details, Transaction List and Details, ForgotPassword, Profile * Created helper function for launching suspended functions on IO thread * Some code refactoring

+++ app/src/main/java/dev/chintansoni/expensetracker/ui/category/dropdown/DropDownView.kt
@@ -2,18 +2,8 @@ package dev.chintansoni.expensetracker.ui.category.dropdown
 
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material.DropdownMenuItem
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.ExposedDropdownMenuBox
-import androidx.compose.material.ExposedDropdownMenuDefaults
-import androidx.compose.material.OutlinedTextField
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
+import androidx.compose.material.*
+import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import dev.chintansoni.domain.model.Category
@@ -23,8 +13,8 @@ import dev.chintansoni.expensetracker.ui.theme.CategoryIcon
 @Composable
 fun RowScope.CategoryView(
     enabled: Boolean = false,
-    selectedCategory: Int = 0,
-    onCategorySelected: (Int) -> Unit = {},
+    selectedCategory: Long = 0,
+    onCategorySelected: (Long) -> Unit = {},
     categories: List<Category>
 ) {
     var expanded by remember { mutableStateOf(false) }
@@ -54,14 +44,14 @@ fun RowScope.CategoryContent(
     expand: Boolean = false,
     enabled: Boolean = false,
     onExpandChange: (Boolean) -> Unit = {},
-    selectedCategory: Int = 0,
-    onCategorySelected: (Int) -> Unit = {},
+    selectedCategory: Long = 0,
+    onCategorySelected: (Long) -> Unit = {},
     categories: List<Category> = emptyList()
 ) {
 
-    val fetchCategoryText: (Int, List<Category>) -> String = { selected, list ->
+    val fetchCategoryText: (Long, List<Category>) -> String = { selected, list ->
         if (list.isNotEmpty()) {
-            if (selected == 0) {
+            if (selected == 0L) {
                 list.first().name
             } else {
                 list.find { it.id == selected }?.name ?: ""
