+++ app/src/main/java/dev/chintansoni/expensetracker/ui/navigator/MainViewRoute.kt
@@ -8,7 +8,7 @@ import dev.chintansoni.expensetracker.ui.auth.splash.ROUTE_SPLASH
 import dev.chintansoni.expensetracker.ui.category.detail.buildCategoryDetailRoute
 import dev.chintansoni.expensetracker.ui.category.list.ROUTE_CATEGORIES
 import dev.chintansoni.expensetracker.ui.home.ROUTE_HOME
-import dev.chintansoni.expensetracker.ui.setting.ROUTE_SETTING
+import dev.chintansoni.expensetracker.ui.home.setting.ROUTE_SETTING
 import dev.chintansoni.expensetracker.ui.transaction.detail.transactionDetailRoute
 
 sealed class MainRoute(
@@ -42,6 +42,6 @@ sealed class MainRoute(
 
     object CategoriesViewRoute : MainRoute(ROUTE_CATEGORIES)
 
-    data class CategoryDetailViewRoute(val categoryId: Int = 0) :
+    data class CategoryDetailViewRoute(val categoryId: Long = 0) :
         MainRoute(buildCategoryDetailRoute(categoryId))
 }
