+++ database/src/main/java/dev/chintansoni/database/view/transactiondetail/TransactionDetailView.kt
@@ -20,7 +20,7 @@ data class TransactionDetailView(
     val id: Long,
     val amount: Float,
     val note: String?,
-    val category: Int,
+    val category: Long,
     val date: Long,
     val createdDate: Long,
     val updatedDate: Long?,
